Restructured Project Structure to be more inline with Redux
